+++ sdk/communication/Azure.Communication.PhoneNumbers/tests/samples/Sample_PhoneNumbersClient.cs
@@ -2,18 +2,11 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure.Communication.PhoneNumbers;
-using Azure.Communication.PhoneNumbers.Models;
-using Azure.Communication.PhoneNumbers.Tests;
-using Azure.Core;
+using Azure.Communication.Tests;
 using Azure.Core.TestFramework;
-using Azure.Identity;
 using NUnit.Framework;
 
 namespace Azure.Communication.PhoneNumbers.Tests.Samples
@@ -34,7 +27,7 @@ namespace Azure.Communication.PhoneNumbers.Tests.Samples
             if (!IncludePhoneNumberLiveTests)
                 Assert.Ignore("Include phone number live tests flag is off.");
 
-            var client = CreateClient(false);
+            var client = CreateClient(AuthMethod.ConnectionString, false);
 
             const string countryCode = "US";
 
@@ -68,7 +61,7 @@ namespace Azure.Communication.PhoneNumbers.Tests.Samples
             var purchasedPhoneNumber = searchOperation.Value.PhoneNumbers.Single();
 
             #region Snippet:UpdateCapabilitiesNumbersAsync
-            var updateCapabilitiesOperation = client.StartUpdateCapabilities(purchasedPhoneNumber, calling: PhoneNumberCapabilityType.Outbound, sms: PhoneNumberCapabilityType.InboundOutbound);
+            var updateCapabilitiesOperation = await client.StartUpdateCapabilitiesAsync(purchasedPhoneNumber, calling: PhoneNumberCapabilityType.Outbound, sms: PhoneNumberCapabilityType.InboundOutbound);
 
             //@@ await updateCapabilitiesOperation.WaitForCompletionAsync();
             /*@@*/ await WaitForCompletionAsync(updateCapabilitiesOperation);
@@ -84,7 +77,7 @@ namespace Azure.Communication.PhoneNumbers.Tests.Samples
             #region Snippet:ReleasePhoneNumbersAsync
 
             //@@var purchasedPhoneNumber = "<purchased_phone_number>";
-            var releaseOperation = client.StartReleasePhoneNumber(purchasedPhoneNumber);
+            var releaseOperation = await client.StartReleasePhoneNumberAsync(purchasedPhoneNumber);
             //@@ await releaseOperation.WaitForCompletionResponseAsync();
             /*@@*/ await WaitForCompletionResponseAsync(releaseOperation);
 
@@ -123,7 +116,7 @@ namespace Azure.Communication.PhoneNumbers.Tests.Samples
             //@@client = new PhoneNumbersClient(endpoint, tokenCredential);
             #endregion Snippet:CreatePhoneNumbersClientWithTokenCredential
 
-            client = CreateClient(false);
+            client = CreateClient(AuthMethod.ConnectionString, false);
 
             const string countryCode = "US";