+++ sdk/communication/Azure.Communication.Sms/tests/SmsClientLiveTests.cs
@@ -172,7 +172,7 @@ namespace Azure.Communication.Sms.Tests
                     InstrumentClientOptions(new SmsClientOptions())));
             try
             {
-                AsyncPageable<SmsSendResult> results = client.SendAsync(
+                var response = await client.SendAsync(
                     from: TestEnvironment.FromPhoneNumber,
                     to: new string[] { TestEnvironment.ToPhoneNumber, TestEnvironment.ToPhoneNumber },
                    message: "Hi",
@@ -180,7 +180,7 @@ namespace Azure.Communication.Sms.Tests
                    {
                        Tag = "marketing", // custom tags
                    });
-                await foreach (SmsSendResult result in results)
+                foreach (SmsSendResult result in response.Value)
                 {
                     Console.WriteLine($"Sms id: {result.MessageId}");
                     assertHappyPath(result);