+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/RecoverDeletedCertificateOperation.cs
@@ -28,6 +28,9 @@ namespace Azure.Security.KeyVault.Certificates
             _response = response.GetRawResponse();
         }
 
+        /// <summary> Initializes a new instance of <see cref="RecoverDeletedCertificateOperation" /> for mocking. </summary>
+        protected RecoverDeletedCertificateOperation() {}
+
         /// <inheritdoc/>
         public override string Id => _value.Id.ToString();