+++ sdk/communication/Azure.Communication.Common/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
-## 1.0.0-beta.6 (Unreleased)
-
+## 1.0.0 (2021-03-29)
+Updated `Azure.Communication.Common` version.
 
 ## 1.0.0-beta.5 (2021-03-09)
 Updated `Azure.Communication.Common` version.